Fix segfault in Damerau-Levenstein C code.
If one of the characters had a value of 128 or above, this would be
treated as a signed char and would result in an array lookup with a
negative index. The somewhat contrived test case given here --
comparing a space with a non-breaking space -- reproduces the
segmentation fault prior to the fix.

This also makes a Clang warning go away. Thanks, compiler! :-)